Expanded links in home page.
Previously for some papers, I had a [more info] link that led to a page
with links to things like the accompanying source code, the tech report,
etc. I realize that people probably don't click through two links as
much as they click on one, so I moved the links in the more info page
into the home page.